server/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Limiter==1.4
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
limits==1.5.1
MarkupSafe==2.0.1
packaging==21.0
proto-plus==1.19.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
rsa==4.7.2
shortuuid==1.0.1
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
google-cloud-translate==3.7.0
